# platform/scenario-b/apis/openapi/amm.yaml
openapi: 3.0.3
info:
  title: CBWeb3 Platform — AMM API (Scenario B)
  version: "2.0.0"
  description: |
    API Gateway endpoints for the Automated Market Maker (Scenario B).
    Covers cooperative liquidity provisioning, commit-reveal, swaps, quotes,
    and multi-pair PairRegistry management (005-cooperative-liquidity).
servers:
  - url: /api/v2/amm
    description: Scenario B AMM namespace
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
paths:
...
  /swap/exact-output:
    post:
      summary: Execute an exact-output swap
      operationId: swapExactOutput
      tags: [Swap]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pair, payer_id, beneficiary_id, amount_out, max_amount_in]
              properties:
                pair:
                  type: string
                  example: "BRL-USD"
                payer_id:
                  type: string
                beneficiary_id:
                  type: string
                amount_out:
                  type: string
                max_amount_in:
                  type: string
      responses:
        "200":
          description: Swap executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_hash:
                    type: string
                  amount_in:
                    type: string
                  order_id:
                    type: string
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Swap failed (POOL_NOT_ACTIVE, CIRCUIT_BREAKER_OPEN, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...